feat(Pagination): extract Pagination from DataTable component
Decouple Pagination from Datatable to allow its standalone use.
