+++ src/components/Pagination/PageButtons.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import { renderWithProviders } from '../../../../../utils/tests/renderWithProviders';
+import { renderWithProviders } from '../../utils/tests/renderWithProviders';
 import PageButtons, { calculatePagePositions } from './PageButtons';
 
 describe('calculatePageButtons', () => {
