+++ src/components/Pagination/PaginationItem.test.tsx
@@ -3,7 +3,7 @@ import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
 import { PaginationItem } from './PaginationItem';
-import { renderWithProviders } from '../../../../../utils/tests/renderWithProviders';
+import { renderWithProviders } from '../../utils/tests/renderWithProviders';
 
 describe('Pagination/PaginationItem', () => {
   it('should not call onClick handler for current page', () => {
